Added copy Game ID to clipboard functionality for easier friends invitation experience

+++ memory-game/src/pages/gameRoom/GameRoom.tsx
@@ -12,6 +12,7 @@ import {
   ListItem,
   Spinner,
   Text,
+  useToast,
 } from "@chakra-ui/react";
 
 export default function GameRoom() {
@@ -22,25 +23,51 @@ export default function GameRoom() {
     "gameRooms",
     gameId
   );
+  const toast = useToast();
 
   const handleLeaveRoom = async () => {
-    const gameRoomRef = doc(db, "gameRooms", gameId);
-    const gameRoomSnap = await getDoc(gameRoomRef);
+    if (gameId) {
+      const gameRoomRef = doc(db, "gameRooms", gameId);
+      const gameRoomSnap = await getDoc(gameRoomRef);
 
-    if (gameRoomSnap.exists()) {
-      const gameRoom = gameRoomSnap.data();
-      if (gameRoom.createdBy.id === user?.uid) {
-        if (gameRoom.opponent) {
-          await updateDoc(gameRoomRef, {
-            createdBy: gameRoom.opponent,
-            opponent: null,
-          });
-        } else {
-          await deleteDoc(gameRoomRef);
+      if (gameRoomSnap.exists()) {
+        const gameRoom = gameRoomSnap.data();
+        if (gameRoom.createdBy.id === user?.uid) {
+          if (gameRoom.opponent) {
+            await updateDoc(gameRoomRef, {
+              createdBy: gameRoom.opponent,
+              opponent: null,
+            });
+          } else {
+            await deleteDoc(gameRoomRef);
+          }
         }
       }
+      navigate("/");
     }
-    navigate("/");
+  };
+
+  const handleCopyToClipboard = async () => {
+    if (gameId) {
+      try {
+        await navigator.clipboard.writeText(gameId);
+        toast({
+          title: "Game ID Copied",
+          description: "Game ID has been copied to clipboard.",
+          status: "success",
+          duration: 2000,
+          isClosable: true,
+        });
+      } catch (err) {
+        toast({
+          title: "Error",
+          description: "Failed to copy Game ID.",
+          status: "error",
+          duration: 2000,
+          isClosable: true,
+        });
+      }
+    }
   };
 
   return (
@@ -64,6 +91,7 @@ export default function GameRoom() {
         <GridItem color="white">
           <Flex flexDir="column" align="center" justify="center" h="100vh">
             <Text>Game ID: {gameId}</Text>
+            <Button onClick={handleCopyToClipboard}>Copy Game ID</Button>
             {roomData?.gameState?.waiting && <Spinner />}
             <Button>Play</Button>
           </Flex>
